It actually works(ish)!!!!!
Lots of hacky workarounds of course. Fixed a lot of things, and it can
now display images and basic stuff. There are also a lot of other
changes that I hadn't committed from PilotDC a month ago, and obviously
I forgot those. So a lot has been changed, definitely.

Today I fixed image viewing and added an Edit function. I also deleted
ImportImageActivity (replaced by ImportImageFragment) and several layout
files that are no longer used.

There are a lot of random Gradle-related changes that are showing up,
and I'm really not sure what to do with them, so I'm not committing
anything that isn't actually part of the code and that I changed.